requirements.txt
Jinja2==3.1.3
MarkupSafe==2.1.5
python-dotenv==1.0.1
PyYAML==6.0.1
